requirements.txt
ultralytics>=8.0.0
opencv-python
numpy
tqdm
pyyaml
scikit-learn
fastapi
uvicorn
python-multipart
pillow
streamlit
